chore: move types in src
